Capstone-Unified/requirements.txt
# =============================================================================
# CRISP UNIFIED PLATFORM REQUIREMENTS
# Combines dependencies from Publication System + Trust Users Management
# =============================================================================

# =============================================================================
# CORE DJANGO FRAMEWORK
# =============================================================================
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
django-extensions>=3.2.0

# =============================================================================
# AUTHENTICATION & SECURITY
# =============================================================================
PyJWT>=2.8.0
cryptography>=41.0.0
argon2-cffi>=23.1.0

# =============================================================================
# DATABASE
# =============================================================================
psycopg2-binary>=2.9.0

# =============================================================================
# THREAT INTELLIGENCE & STIX PROCESSING (Publication System)
# =============================================================================
stix2>=3.0.0
taxii2-client>=2.3.0
requests>=2.31.0
lxml>=4.9.0
xmltodict>=0.13.0
jsonschema>=4.19.0

# =============================================================================
# BACKGROUND PROCESSING (Publication System)
# =============================================================================
celery>=5.3.0
redis>=4.5.0
django-redis>=5.3.0

# =============================================================================
# API DOCUMENTATION
# =============================================================================
drf-spectacular>=0.26.0

# =============================================================================
# DEVELOPMENT & TESTING
# =============================================================================
coverage>=7.0.0
factory-boy>=3.3.0
faker>=19.0.0
python-dotenv>=1.0.0

# =============================================================================
# PRODUCTION DEPLOYMENT
# =============================================================================
gunicorn>=21.0.0
whitenoise>=6.5.0

# =============================================================================
# DATA PROCESSING & UTILITIES
# =============================================================================
python-dateutil>=2.8.0
pytz>=2023.3
validators>=0.22.0
pyyaml>=6.0.1

# =============================================================================
# EMAIL & COMMUNICATION
# =============================================================================
# Using Django's built-in SMTP backend - no additional packages needed

# =============================================================================
# MONITORING & PROGRESS TRACKING
# =============================================================================
tqdm>=4.66.0

# =============================================================================
# VERSION CONSTRAINTS FOR COMPATIBILITY
# =============================================================================
setuptools>=68.0.0
wheel>=0.41.0

# =============================================================================
# OPTIONAL DEVELOPMENT TOOLS (uncomment as needed)
# =============================================================================
# django-debug-toolbar>=4.1.0
# ipython>=8.14.0
# jupyter>=1.0.0faker>=19.0.0
tqdm>=4.65.0
